docker-apps/nomad-dashboard/exporter/requirements.txt
prometheus-client==0.19.0
requests==2.31.0
python-dotenv==1.0.0 